// frontend/controller/interfaces/Project.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Project {
  id: number
  type: string
  url: string
  related: {
    created_by: string
    modified_by: string
    teams: string
    playbooks: string
    inventory_files: string
    update: string
    project_updates: string
    scm_inventory_sources: string
    schedules: string
    activity_stream: string
    notification_templates_started: string
    notification_templates_success: string
    notification_templates_error: string
    access_list: string
    object_roles: string
    copy: string
    organization: string
    last_job?: string
    last_update?: string
  }
  summary_fields: {
    organization: {
      id: number
      name: string
      description: string
    }
    created_by: {
      id: number
      username: string
      first_name: string
      last_name: string
    }
    modified_by: {
      id: number
      username: string
      first_name: string
      last_name: string
    }
    object_roles: {
      admin_role: {
        description: string
        name: string
        id: number
      }
      use_role: {
        description: string
        name: string
        id: number
      }
      update_role: {
        description: string
        name: string
        id: number
      }
      read_role: {
        description: string
        name: string
        id: number
      }
    }
    user_capabilities: {
      edit: boolean
      delete: boolean
      start: boolean
      schedule: boolean
      copy: boolean
    }
    last_job?: {
      id: number
      name: string
      description: string
      finished: string
      status: string
      failed: boolean
    }
    last_update?: {
      id: number
      name: string
      description: string
      status: string
      failed: boolean
    }
  }
  created: string
  modified: string
  name: string
  description: string
  local_path: string
  scm_type: string
  scm_url: string
  scm_branch: string
  scm_refspec: string
  scm_clean: boolean
  scm_track_submodules: boolean
  scm_delete_on_update: boolean
  credential: null
  timeout: number
  scm_revision: string
  last_job_run: null | string
  last_job_failed: boolean
  next_job_run: null
  status: string
  organization: number
  scm_update_on_launch: boolean
  scm_update_cache_timeout: number
  allow_override: boolean
  custom_virtualenv: null
  default_environment: null
  signature_validation_credential: null
  last_update_failed: boolean
  last_updated: null | string
}